requirements.txt
click==8.1.1
colorama==0.4.4
Flask==2.1.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.23
Werkzeug==2.1.0
zipp==3.7.0
